refactored login activity & added dependency injection
